Add a proper WindowProperties object
The original "named properties tracker" was created before Proxy was available. Now we can implement this in a way much closer to the spec.

As part of this, restructure Window.js to be less convoluted, with a single createWindow function and no fake Window constructor that operates on a this object but then manually sets is prototype chain anyway.